moved project_plot2d to examples
